Dostosuj gałąź do ESLint'a
