+++ src/screens/Groups/Groups.tsx
@@ -1,57 +1,17 @@
-import React, {useCallback, useState} from 'react';
-import {FlatList, StyleSheet, View} from 'react-native';
-import {DataWithIsChecked} from "./index";
-import GroupListItem from "./GroupListItem";
-import EmptyListComponent from "./EmptyListComponent";
-import {FAB} from "react-native-paper";
-import AddGroupModal from "./AddGroupModal";
+import React, { useCallback, useState } from 'react';
+import { FlatList, StyleSheet, View } from 'react-native';
+import { DataWithIsChecked } from './index';
+import GroupListItem from './components/GroupListItem';
+import EmptyListComponent from './components/EmptyListComponent';
+import { FAB } from 'react-native-paper';
+import AddGroupModal from './components/AddGroupModal';
 
 interface Props {
-    data: DataWithIsChecked[],
-    onGroupPress: (groupId: number, isIncluded: boolean) => void,
-    onLongGroupPress: (groupId: number) => void,
-    addGroup: (name: string) => void,
+    data: DataWithIsChecked[];
+    onGroupPress: (groupId: number, isIncluded: boolean) => void;
+    onLongGroupPress: (groupId: number) => void;
+    addGroup: (name: string) => void;
 }
-
-const Groups = (props: Props) => {
-    const [modalVisible, setModalVisible] = useState(false);
-    const {data, onGroupPress, addGroup, onLongGroupPress} = props;
-
-    const renderItem = useCallback(
-        ({item}) => (
-            <GroupListItem
-                item={item}
-                onGroupPress={onGroupPress}
-                onLongGroupPress={onLongGroupPress}
-            />
-        ),
-        []);
-
-    return (
-        <>
-            <AddGroupModal modalVisible={modalVisible} setModalVisible={setModalVisible} addGroup={addGroup}/>
-            <View style={styles.container}>
-                <FlatList
-                    data={data}
-                    numColumns={1}
-                    contentContainerStyle={styles.listContentContainer}
-                    renderItem={renderItem}
-                    keyExtractor={(item) => item.id.toString()}
-                    ListEmptyComponent={<EmptyListComponent/>}
-                />
-                <FAB
-                    visible={!modalVisible}
-                    style={styles.fab}
-                    icon="plus"
-                    onPress={() => {
-                        setModalVisible(true);
-                    }}
-                />
-
-            </View>
-        </>
-    );
-};
 
 const styles = StyleSheet.create({
     container: {
@@ -69,5 +29,49 @@ const styles = StyleSheet.create({
         zIndex: 200,
     },
 });
+
+const Groups = (props: Props) => {
+    const [modalVisible, setModalVisible] = useState(false);
+    const { data, onGroupPress, addGroup, onLongGroupPress } = props;
+
+    const renderItem = useCallback(
+        ({ item }) => (
+            <GroupListItem
+                item={item}
+                onGroupPress={onGroupPress}
+                onLongGroupPress={onLongGroupPress}
+            />
+        ),
+        [onGroupPress, onLongGroupPress],
+    );
+
+    return (
+        <>
+            <AddGroupModal
+                modalVisible={modalVisible}
+                setModalVisible={setModalVisible}
+                addGroup={addGroup}
+            />
+            <View style={styles.container}>
+                <FlatList
+                    data={data}
+                    numColumns={1}
+                    contentContainerStyle={styles.listContentContainer}
+                    renderItem={renderItem}
+                    keyExtractor={(item): string => item.id.toString()}
+                    ListEmptyComponent={<EmptyListComponent />}
+                />
+                <FAB
+                    visible={!modalVisible}
+                    style={styles.fab}
+                    icon='plus'
+                    onPress={(): void => {
+                        setModalVisible(true);
+                    }}
+                />
+            </View>
+        </>
+    );
+};
 
 export default Groups;
