+++ src/screens/Groups/components/EmptyListComponent.tsx
@@ -1,0 +1,28 @@
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+import { MaterialCommunityIcons } from 'react-native-vector-icons';
+
+const styles = StyleSheet.create({
+    informationContainer: {
+        flex: 1,
+        marginBottom: 3,
+        margin: 5,
+        justifyContent: 'center',
+        alignItems: 'center',
+    },
+    message: {
+        fontSize: 18,
+        margin: 5,
+    },
+});
+
+const EmptyListComponent = () => {
+    return (
+        <View style={styles.informationContainer}>
+            <MaterialCommunityIcons size={100} name={'account-group'} />
+            <Text style={styles.message}>No groups available. Add one!</Text>
+        </View>
+    );
+};
+
+export default EmptyListComponent;
