+++ src/screens/Groups/index.tsx
@@ -1,66 +1,77 @@
-import React, {useMemo} from 'react';
-import {useDispatch, useSelector} from 'react-redux';
-import {getGroups} from '../../redux/selectors/Selectors';
-import GroupsView from "./Groups";
-import {addContactToGroup, createGroup, removeContactFromGroup, removeGroup} from "../../redux/actions/ActionCreators";
-import {GroupWithoutId} from "../../redux/reducers/GroupsReducer";
-import {Alert} from "react-native";
+import React, { useMemo } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import { getGroups } from '../../redux/selectors/Selectors';
+import GroupsView from './Groups';
+import {
+    addContactToGroup,
+    createGroup,
+    removeContactFromGroup,
+    removeGroup,
+} from '../../redux/actions/ActionCreators';
+import { Group } from '../../redux/reducers/GroupsReducer';
+import { Alert } from 'react-native';
 
 interface Props {
-    route: any
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    route: any;
 }
 
 export interface DataWithIsChecked {
-    id: number,
-    name: string,
-    contactsIds: number[],
-    isChecked: boolean,
+    id: number;
+    name: string;
+    contactsIds: number[];
+    isChecked: boolean;
 }
 
 const GroupsScreen = (props: Props) => {
-    const {id} = props.route.params;
+    const { id } = props.route.params;
     const groups = useSelector(getGroups);
     const dispatch = useDispatch();
 
-    const onGroupPress = (groupId: number, isIncluded: boolean) => {
-        isIncluded ? dispatch(removeContactFromGroup(id, groupId)) : dispatch(addContactToGroup(id, groupId));
+    const onGroupPress = (groupId: number, isIncluded: boolean): void => {
+        isIncluded
+            ? dispatch(removeContactFromGroup(id, groupId))
+            : dispatch(addContactToGroup(id, groupId));
     };
 
-    const addGroup = (name: string) => {
+    const addGroup = (name: string): void => {
         if (!name) return;
-        let group: GroupWithoutId = {name: name, contactsIds: []};
+        const group: Group = { id: null, name: name, contactsIds: [] };
         dispatch(createGroup(group));
     };
 
-    const onLongGroupPress = (groupId: number) => {
+    const onLongGroupPress = (groupId: number): void => {
         if (!groupId) return;
         Alert.alert(
-            "Group",
-            "Do you want to delete this group permanently?",
+            'Group',
+            'Do you want to delete this group permanently?',
             [
                 {
-                    text: "Cancel",
-                    onPress: () => {
-                    },
-                    style: "cancel"
+                    text: 'Cancel',
+                    style: 'cancel',
                 },
-                {text: "OK", onPress: () => dispatch(removeGroup(groupId))}
+                { text: 'OK', onPress: (): void => dispatch(removeGroup(groupId)) },
             ],
-            {cancelable: true}
+            { cancelable: true },
         );
     };
 
     const data: DataWithIsChecked[] = useMemo(() => {
-        return groups.map((el, i) => {
+        return groups.map((el) => {
             return {
                 ...el,
                 isChecked: el.contactsIds.includes(id),
             };
         });
-    }, [groups]);
+    }, [groups, id]);
 
     return (
-        <GroupsView data={data} onGroupPress={onGroupPress} onLongGroupPress={onLongGroupPress} addGroup={addGroup}/>
+        <GroupsView
+            data={data}
+            onGroupPress={onGroupPress}
+            onLongGroupPress={onLongGroupPress}
+            addGroup={addGroup}
+        />
     );
 };
